requirements.txt
langchain>=0.2.0,<0.3.0
langchain-community>=0.2.0,<0.3.0
langgraph>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
streamlit>=1.56.0
python-dotenv>=1.0.0
pydantic>=2.11.0,<2.13.0
loguru>=0.7.0
redis>=5.0.0
rank-bm25>=0.2.0
FlagEmbedding>=1.2.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
transformers>=4.43.0
torch>=2.0.0
zhipuai>=2.0.0